+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - alow to set height per minute in [InRowCalendarDayView] and [OverflowItemBuilder]
 - option to show a line that indicates current hour and minute in day view
+- refactor code to reflect height per minute correctly
 
 ## 1.0.2